2894160 by jochemvn, Kingdutch: implement Entity::getDisplayName for posts
We implement a new method for entities that is already implemented
by User. This pushes the logic of how to display something into the
Entity itself.

This new method is used by notifications and falls back to the
current behaviour of the bundle. This improves extensibility and
avoids the necessity for more special cases when bundle names of
new entity types are not suitable for display to the user.

For now the implementation on the Post entity type will use the
label of the Entity (which is Post) but we could implement some
logic there to use "Photo" if the foto field is filled and "Post"
otherwise.