Improve margin targetting of comment paragraphs
It looks better if the like button is flush against the bottom of
a comment. To achieve this the margin-bottom of paragraphs in
comment text is set to 0. This has the drawback that long
paragraphs become very difficult to read.

This commit changes the targetting of the margin-bottom to be only
for the last paragraph in a comment. This ensures that for both
single paragraph and multi-paragraph comments the like button is
flush against the last paragraph. However it ensures that spacing
between multiple paragraphs is retained.

This changes nothing for short, one paragraph comments.