28494160 by Kingdutch: Fix no such field error for posts
When a post is created in a test using the 'post' bundle instead
of the 'image' bundle (which is what users in generally create in
the stream) then there is no field called 'field_post_image' to
test for. Therefor we first need to test if the field exists before
deciding it has a value.

We test for the field existance instead of the bundle because we
can't trust the bundle and this error might pop up if a bundle is
added without image field but with a name other than 'post'.