Issue #2909681 by Kingdutch: Call parent::preprocess for socialbase alter plugins
We dictate that subthemes extend the SocialBase plugins so that
our functionality at least gets all the required theming variables
it requires. The function `PreprocessBase::process` in the
Bootstrap theme does quite some heavylifting for us that a
subtheme plugin might depend on. We should therefor ensure that
this function is always called whenever we overwrite the function
in our socialbase theme.