Alter the requirements for the Dropdown::isEmpty check
Nowhere in the Dropdown do we require that the value that is
supplied is actually an integer. This causes some implementations
where we pass in a string '0' instead. To account for htis we
change our checks to see where we either have no value at all or
where we have a value but it's neither an integer nor a string.