Issue #2927442 by Kingdutch: fix undefined index in social_profile_privacy.install
Instead of throwing an error we return an empty array for roles
that we don't want to explicitly grant permissions to. The users
with these roles will also inherit the permissions from the
"authenticated user" role.

This also removes the need for a special case for the administrator
role.

Passing an empty array to `user_role_grant_permissions` will not
modify any permissions.