Issue #2944720 by Kingdutch: Profile anonymous theme suggestion is too important
The following commit that introduced Public Group support added a theme hook for profiles that is used when anonymous users view the page: https://github.com/goalgorilla/open_social/commit/7d539a7120299e7f9f9804.... However, because that suggestion is simply added to the stack it is used over any other custom profile displays that may be used. This breaks existing sites.

The solution to this is to add the suggestion only after the suggestion it supplements. This way the original importance is maintained and any anonymous template takes precedence over it's non-anonymous counterpart.

In a solution for this I've extended the hooks to add anonymous versions for all suggestions that are implemented by the profile module which provides some additional flexibility.

This results in the following list of file name suggestions:

```
<!-- FILE NAME SUGGESTIONS:
   * profile--305--teaser--anonymous.html.twig
   * profile--305--teaser.html.twig
   * profile--305--anonymous.html.twig
   * profile--305.html.twig
   x profile--profile--teaser--anonymous.html.twig
   * profile--profile--teaser.html.twig
   * profile--profile--anonymous.html.twig
   * profile--profile.html.twig
   * profile--teaser--anonymous.html.twig
   * profile--teaser.html.twig
   * profile--anonymous.html.twig
   * profile.html.twig
-->
```