Issue #2916984 by chmez, Kingdutch: Fix deleting post comment. (#587)
* Issue #2916984 by chmez: Fix deleting post comment.

* Issue #2916984 by chmez, Kingdutch: Fix deleting post comment.

This commit is guards slightly better against comment bundles that
don't have a field_comment_files field. The get magic method has
the side effect of adding a NULL value in the values array and
thus changes the entity.

* Issue #2916984 by chmez, Kingdutch: Fix deleting post comment.

This commit moves the file cleanup from the comment_delete
hook to a separate function. This allows a simple guard hook to
be used so that no errors are caused when a post without file fields
is deleted. It also prevents the update hook not being executed
at all if this guard clause is forgotten when code is added in
the future.

See https://github.com/goalgorilla/open_social/pull/587#issuecomment-342510266
for the initial discussion.