Don't use existing configuration in overrides
Configuration overrides are done using a deep merge which means
that only keys existing in the overrides array are overwritten with
as much specificity as possible. We therefor don't need the entire
`content` array but only the field of the `content` array that we
want to add.

By loading the entire `content` array first we increase the amount
of configuration marked as 'overridden' which is then more difficult
for other modules to alter.