OSSUPPORT-385 Enable content translation for page types
This creates a dummy module that is used as the dependency for the
content translation settings of the Landing Page, Book Page and
Basic Page content types. This way the content translation can be
easily enabled and disabled by enabling the module.

Content translation configuration is stored with the modules that
control the content type's fields so that they can be updated when
needed.

Translation for the paragraphs is also enabled because they make
up a large part of the landing page content.