OSSUPPORT-385 Implement sane content translation defaults
This groups the sane defaults for translation of basic pages,
landing pages and book pages. These defaults are only applied
when the social_language_content module is enabled and the setting
for that content type is set to true. This allows another module
(for example a UI) to determine which content types are
translatable for a sitemanager.

These defaults are applied as a dynamical config override because
they're contained within other .yml files that we do not want to
edit because it would create a new dependency (or break without
the content language module enabled).