OSSUPPORT-385 Disable content translation for all fields by default
Drupal's default behaviour is to enable content translation for
all fields if we enable it for a certain entity bundle. However,
our config overrides assume that it's instead disabled for all
fields and we enable it selectively.

To ensure we get this described behaviour we add an install hook
that disables the translations for all the entities that we
currently support.