Issue #2956238 by Brhane, Kingdutch: Incorrect "event has passed" message
When changing a date from a string to a time PHP uses the timezone
that's returned by `date_default_timezone_get`. However, this is
set by Drupal to be set to the timezone of the user that is making
the current request. This causes a time that is stored as UTC to
be treated in a different timezone. If a `time()` call is then used
to compare against this timezone then in timezones that are behind
`UTC` they'll be able to enroll to events that have passed while
in timezones ahead of `UTC` events will close early.

DateTime fields also have a `date` property which stores the time-
zone aware value. We then use `new DrupalDateTime` which defaults
to `time()` in server time and creates another proper timezone
aware value. Those values can then be compared and the values will
be modified so that comparison takes place in the same timezone.

This results in the proper timezone aware closing of events where
the event enrollment closes for all users around the globe at the
same moment.