Issue #2968895 by Kingdutch: Improve activity_send_email_form_user_form_alter code quality
The if statement that was here is not needed because in the user
form the account object is always present. The new code more
properly displays the intent of the code to exclude the settings
form from user creation (or registration) forms. This cahnge also
decouples this alter hook from the Drupal routing service.