Write a test for the 404 landing page bug
This test should fail since the fix will be in a later commit. It
tests that a landing page used as a 404 page will incorrectly
display the page title block.

The page title block should not be displayed above landing pages
used as 404s or 403s because these landing pages have their own
heroes.