OSSUPPORT-615 by Kingdutch: Fix landing page title as 404 page
This commit adds some logic to the PageTitleBlock to figure out if
it's displayed on a landing page as part of a 404 or 403 response
subrequest. When a landing page is used as one of these error pages
then the title block is not needed because landing pages have their
own ways of creating heroes.