Convert old template array to a Drupal core render array
We transform the render array that was used for the custom template
file to one that can be used with mostly Drupal core render elements.

We add one UnwrappedContainer render element that provides us with
a container without a div. This ensures that we can have multiple
child elements in a <li> element without the list_item element
rendering multiple <li> elements.

This provides a good basis for simplifying it into our own render
element.