Issue #2979293 by Kingdutch: Change download hover text
Wether a file opens in the browser or is downloaded depends on the
file format and the browser configuration (type, addons, etc) of
the user. Currently the text "Open file in a new window" is used
which gives a user the expectation of seeing the file in a new
window even if this is not necessarily the case.

We change the text to prime the user for the possible behaviours
of the browser.