Issue #2941556 by Kingdutch, jaapjan, bramtenhove: Editing posts in a closed group impossible due to issue with the visibility form element
Refactor the PostForm form to more clearly distinguish the various
parts of its build steps. Also try to add comments to denote the
process that figures out which visibility modes may be selected.

Fixes the bug where in edit mode there's no profile or group to be
gleaned from the URL. It also ensures that the same view mode is
used for creating a Post (which is set by the block rendering the
form) as well as editing a post (where the form is rendered by
Drupal).