DS-3488 by bramtenhove: Replace Core's AdminCommentOverview page.
Replaced the default overview to be able to strip out the 'Subject' and 'Posted on' columns. Instead we now have the actual comment visible with extra operations to make it easier to manage the comments.

The comment body is rendered in full display mode so links and mentions also work.