Move the links from the AccountHeaderBlock to their respective modules
This commit moves the links from the AccountHeaderBlock for the user
profile dropdown menu to the modules that they belong to.

Where a module belongs is defined by the route it points ot and
the module that manages that route. This ensures that if the
functionality of a module is disabled and the route no longer
exists then the link will also not exit. This helps in the
decoupling of the various modules.