Write compatibility layer for deprecated alter hook
The alter hook that was previously supported to add links to the
account menu must remain supported in the new system. We implement
the new alter hook that calls the old one and tries to properly
insert any links that may have been provided.

To make sure that arrays don't need to be regenerated multiple
times it traverses the arrays and ensures that weights are
present.