Fix case sensitivity of "New content" links
In the account header block all the create content links are in
the form "New Event" or "New Group" where the second word is
also capitalized. The test does not have this capitalization which
causes the translation to fail (because it's case sensitive). This
is fixed by translating the correct casing of the words.