Issue #2994272 by Kingdutch: Increase waiting time for AJAX requests in Behat tests
Paragraphs can be quite slow to load by AJAX. Especially with a lot of elements such as what happens in some landing page tests. This can cause the 5 second timeout for the iWaitForAJAX step to be insufficient.

To solve this we extend the time to 20 seconds which should be enough to let the AJAX requests pass even in slower conditions. Subjective tests show it takes about 10-15 seconds so this should provide a safe margin.