Remove the parent for a post_comment type comment
Comments on posts can not be a reply to another comment on a post.
Thus a comment on a post can never have a parent. The demo content
contained a comment for a post with a parent which caused a hard
to debug bug in the social_mentions module that was discovered while
working on issue #3003810.