Use proper English texts for the AN Event Enrollment settings form
This is the result of the code review that was done by Natasha.

We ended up on using `event organiser(s)` because this was in line
with the functionality that could be provided by the Event Organiser
module. It was most fitting for the first setting and was then
consistently applied to the other settings.