Provide a method for Open Social installations to fill-in custom form fields
Platforms that are based on Open Social but have more than just the
core distribution often add custom fields to Entities that are
required. This causes an issue when the platform tries to use the
Open Social tests to verify that basic functions still work because
those required fields cause validation errors when creating test
content.

To solve this we add a dummy step that those platforms can overwrite
to fill in those custom values for the distro tests which should
lead to a reduction of duplicated tests.