Change wording of "anonymous enrollments" for platform users
After a remark from one of our customers we've discussed what to do
with the texts in the anonymous enrollment feature. We have decided
to continue using the term "anonymous enrollment" and "anonymous
user" for our settings page because it corresponds to documentation
and is in line with other screens that are available to the site
manager.

For 'regular' platform users we've gone with "Sign up as Guest"
for people that are not already a member of the platform. When
creating an event we've chosen the term "Allow users to enroll for
this event without creating an account" because using the phrasing
"Allow users to enroll as guest" may confuse the user into thinking
that the checkbox controls all user enrollment for that event.

The tab for "Anonymous enrollment" has been renamed to "Guest
enrollment" because the enrollments are not anonymous (people fill
out user info) but they are guests to the platform.

The warning message for when anonymous enrollment is enabled for
an event but the visibility of the event is changed from public to
something else has been revised to be in line with the wording of
the checkbox itself: "without account".