Refactor LanguageSwitcherBlock to use core LanguageSwitcher
This refactors the language switcher block to reduce the amount of
code and fix an issue with prefixes. It also adds the block to both
socialbase and socialblue when the social_language module is enabled.

The block should ensure that it renders nothing if there's only a
single language for the user.

Optionally the block and optional config should be moved back to
the social_content_translation module to not alter the behaviour of
the social_language module.