Control entity translatbility in translation override classes
The translate overview for entities is shown depending on whether
the entity is translatable. This means that if an entity is
translatable but none of its fields are then the translation overview
(e.g. node/*/translations) is still shown even though a user can't
really do anything with it.

To solve this issue the entity translation is disabled by default
when the social_content_translation module is enabled and the
specific content type override then ensures that it is made
translatable.