Fix content translation override cacheability
If the configuration for the social_content_translation module
changes then the configuration overrides for content translation
need to be updated because some content types may no longer be
translatable.

This commit alters how the configuration overrides are stored
within the class so that the cache tags for cacheability can be
added at the right moment (when an overwritten configuration is
requested).