Allow request static cache to be disabled for SocialGroupHelperService::getGroupFromEntity
The static cache can cause issues in requests where the group that
a node belongs to is changed during the request. This happens for
example in GroupActivityContext::getRecipients which is called
when a node is moved between groups. To fix this we allow the read
part of the cache to be ignored (optionally). This means that by
default the function is performant but call-sites that know that
the value has changed can invalidate the cache.

We allow the read-only value to be ignored and still update the
value to ensure that in the above scenario the next call-site that
may be unaware the value has changed still receives the updated
value instead of the old value.