Replace normal URLs by URL::toRenderable in AccountHeaderBlock links
The URL::toRenderable will add the access check method for the URL to
the render array. This can then be used by the link to determine whether
it should render or not. If we only add the URL under the #url key then
unfortunately this access check does not happen.