Export missing translations
Some of the translations were missing due to the way we used the potx
tool. The usage has been adjusted to provide the tool with more context
about the schemas within Drupal and the modules that Open Social uses.
This also means that we no longer require a custom yml file for the potx
tool.