Fix project names in fully qualified dependencies
When using the project:module_name syntax for dependency declaration the
`project` part should reflect the name under which the project can be
found on drupal.org. For all modules contained in Open Social this is
thus the `social` project.

This change allows automated systems (e.g. automatic updates or
translation servers) to find the correct module even if a module is
contained in multiple projects.