Fix styling for social sharing buttons
The Shariff buttons don't have rounded corners or centered icons by
default which causes a regression compared to the previously present
addtoany buttons.

This commit adds a default border-radius for the socialblue theme and
uses the socialbase theme to center the icons. It also ensures that when
a theme sets a custom border-radius for buttons that it's honored on the
sharing buttons.