Add proper code linting using eslint and prettier
We set up ESLint to lint our code using the AirBnB styleguide which is
full of good practices. We also use prettier for more opiniated styles
and autofixes and then wire the two together.

We make some changes to our webpack build because prettier doesn't allow
extensions in import statements (which makes sense because they're an
implementation detail). This allows webpack to properly find .jsx files
as well (because jsx is not allowed in a .js file).