Set-up React rendering under the search block
Adjusts the styling for search to work when more items such as the
search suggestions field are added. Also adds weights to the search form
to ensure that it can be properly sorted without rearranging its
contents.

Finally adds the compiled module javascript as a library and attaches it
to the search form.