Move React library to a separate module
We shouldn't bundle React code together with components that use React
as this could cause the library to be loaded multiple times on a single
page.

React is moved to a separate module so that other modules that are
created in the future can rely on this module as well and does not need
to add its own copy of React.