Connect React component to Search REST endpoint
This commit changes the REST endpoint to normalize its output for the
entity types that we have. The URL data that is needed is passed to the
React components through Drupal.settings. Finally jQuery is used to
fetch results from this REST endpoint and render them in our components.

At this point it is not yet possible to pull the enrollment data for
events into the REST endpoint so those tags can not yet be displayed.

The only thing to improve UX is to add a progress bar during the AJAX
request.