Highlight all words in the search query
Highlighting would stop working when a string was given that didn't
match exactly in the search result (e.g. "open social" would be
highlighted in "Open Social has it's birthday" but "Open Evolving" would
not be highlighted in "Open Social is still evolving"). This commit
ensures that all words in the search query will be highlighted if
they're present.