Cancel previous search result request
It's weird for the user if they keep getting changing suggestions. It
makes more sense to cancel the previous request to the server when
requesting new data. This will just keep the loader visible for users
instead.

It's not known whether this actually has effect on server performance
because most resources seem to say that the server will just continue
what it's doing. It'll be up to Drupal to figure out the request has
terminated and stop processing.