Final design tweaks for autosuggestions
After a short meeting with Xinyu we decided to reduce the font-size for
the node type to de-emphasize this and let the user focus on the main
content. The spinner has also been changed to follow the brand colors.
In this case it follows the navigation text color when displayed in the
search overlay.