Add report as topic link
We move the moderation links into a subgroup and kill the lazy_builder.
This allows the moderation links (currently only 'report') to be
rendered as a sub-group in the node template. This allows the sub-group
to be rendered using a separate template file which creates the dropdown
menu we need.