Fix type display for user entities
The type for User entities was not properly displayed in search. It
appears that the user entity type is only visible to the admin user. By
rewriting the uid to the word "User" we fix this. The uid will only be
set (and thus rewritten) when the loaded entity is a user entity.