Fix html entities doubly escaped in search suggestions
This fixes an issue where the output in the search suggestions was
escaped twice. This is caused by the REST API which escapes its output
within the JSON encoding. This is currently an open Drupal core bug. To
work around this we decode the entities first when receiving the
response and then just let React do its normal escaping.

See https://www.drupal.org/project/drupal/issues/2928793