Use profile entity from form in edit form alter
In our code we're always dealing with the user whom the profile belongs
to and this is taken care of for us by the form. There's no use in
trying to retrieve this same information from the URL.

In addition this commit doesn't load the User entity to whom the profile
belongs but instead just asks the profile for the ID directly.

Finally it renames the generic `$user` variable to be less ambigious.