Display profile tags in a hierarchy in search
The profile tag filter in the user view is now displayed hierarchically
if the taxonomy contains multiple levels of items. This corresponds to
the changes made on the profile. Functionally there's no difference in
the search yet.

This change also causes the ordering of the taxonomy to be respected in
the view (there was a setting for that).

If you're viewing this because the preprocess hooks are causing
performance issues: I'm sorry, I'm not proud. Good luck!