Remove geolocation code for search block
This code was added long ago in 09319ab but the module has since been
removed from Open Social and used in another project. In this other
project the block that was referenced to has been removed because it was
no longer needed. Finally the module was re-introduced as a stand-alone
extension when it was extracted from the project. This has surfaced the
bug that this code now causes.