Issue #2946146 by Kingdutch, sukh.singh: Don't run entity queries for activities during configuration entity modification.
Activities should never be created for config entities (they notify of
content). However, in certain cases (feature reverts, module uninstall,
update hooks) it can cause issues when an entity query is executed for
activities during a configuration change. The easiest way is to just
stop doing things when we're triggered for a configuration entity.