Move first name and last name into search indices
This change moves the first name and last name of user profiles from
being indexed through the rendered entity into their own fields in the
index. This allows us to apply access control to the fields which is a
new feature that will be implemented in the social_profile_privacy
module.