Move full name protection code to social_profile_privacy
This moves the code for hiding the full name of users to the
social_profile_privacy. It also modifies the `useFullName` condition to
work regardless of there being a nickname. This ensures that even if the
social_profile_fields module is not enabled then the Full Name can still
be hidden and the username used as display name.

The only issue remaining in this commit is that the full name hiding
happens in the social profile fields module. This will need a better
solution by moving the authoritative rule on this to the social user
module.