Add user display name resolution to social_user
Multiple modules were trying to overwrite the user display name. The
social_profile module wanted to display the user's full name. The
social_profile_fields module wanted a nickname to take precedence and
finally the social_profile_privacy wanted a say in when the full name
would be displayed.

To resolve this we centralise the username resolution in the social_user
module and provide other modules the ability to add suggestions using a
weight to specify order of preference. An alter hook is implemented to
give modules the chance to overwrite or unset suggestions. This can be
used for example to implement access checking.

If no suggestions are left then we simply do not alter the name and
default to Drupal's built-in behaviour.