Ensure search content is indexed in tests
In f9ff903 we decided that search content would no longer be indexed
immediately in order to improve Open Social performance. However, tests
rely on content being indexed during testing. To fix this a test step
has been added to trigger search indexing. This also allows the scenario
of content not being indexed to be tested should a user so desire.