Disable report overview SQL rewriting
There is an issue that causes reports for things other than nodes not to
show up.

This seems to be caused by the following Drupal core issue: https://www.drupal.org/project/drupal/issues/1349080

The short summary is that node access rules are applied to the main query (loading the flags) instead of only the node subquery. This causes any data without node grants (comments and posts) not to match anything and return no results.

We've (Bram ten Hove and I) decided to disable SQL rewriting in the view for now. This fixes the issue but may cause us trouble later. Once the Drupal issue is fixed then we can re-enable SQL rewriting and everything should still work.