Issue #3045199: Implement EntityOwnerInterface for MentionsInterface
The Mentions entity already implemented some of the EntityOwnerInterface
methods and places using this entity also relied on those methods being
present. This change ensures that things that say they are a compatible
with usages of the Mentions entity, actually are.

It also implements the setOwnerId and setOwner methods as convenience
methods in order to adhere to the Interface.