Issue #3039084 by chmez, Kingdutch, jochemvn, jaapjan: Allow full name to be hidden (#1255)
* Issue #3039084 by chmez: Create permission for always showing Full Name.

* Issue #3039084 by chmez: Add new option on the settings page.

* Issue #3039084 by chmez: Check if can using a full name for search for mentions.

* Issue #3039084 by chmez: Use shorter format for conditions for searching in a part of a field value.

* Issue #3039084 by chmez: Turn off limitation by default.

* Issue #3039084 by chmez: Fix error when searching users when profile tags are absent.

* Issue #3039084 by chmez: Prohibit using the full name for search by users without permission when the limitation is turned on.

* Issue #3039084 by chmez: Extend description for name argument of sorting method and delete sorting method name which using by default.

* Issue #3039084 by chmez, kingdutch: Use longer name of configuration.

* Issue #3039084 by chmez: Fix getting username instead of nickname when the full name is empty.

* Issue #3039084 by chmez: Use single-line format for method for checking if can using full name.

* Add nickname field to search_all index

This is needed so the nickname can be used in search filter queries that
will be implemented for the privacy module. This should also improve
search_all search results.

* Move first name and last name into search indices

This change moves the first name and last name of user profiles from
being indexed through the rendered entity into their own fields in the
index. This allows us to apply access control to the fields which is a
new feature that will be implemented in the social_profile_privacy
module.

* Fix preprocessing for name fields in search

The fields need to be preprocessed so that they are properly tokenised
and work with non-ascii characters.

* Filter out matches on full name for non-eligible users

When the users do not have the permission to see the full name and the
full name is protected then the matches in search that arrived from the
name and no other field are filtered out.

* Move full name protection code to social_profile_privacy

This moves the code for hiding the full name of users to the
social_profile_privacy. It also modifies the `useFullName` condition to
work regardless of there being a nickname. This ensures that even if the
social_profile_fields module is not enabled then the Full Name can still
be hidden and the username used as display name.

The only issue remaining in this commit is that the full name hiding
happens in the social profile fields module. This will need a better
solution by moving the authoritative rule on this to the social user
module.

* Add user display name resolution to social_user

Multiple modules were trying to overwrite the user display name. The
social_profile module wanted to display the user's full name. The
social_profile_fields module wanted a nickname to take precedence and
finally the social_profile_privacy wanted a say in when the full name
would be displayed.

To resolve this we centralise the username resolution in the social_user
module and provide other modules the ability to add suggestions using a
weight to specify order of preference. An alter hook is implemented to
give modules the chance to overwrite or unset suggestions. This can be
used for example to implement access checking.

If no suggestions are left then we simply do not alter the name and
default to Drupal's built-in behaviour.

* Update full name hiding texts and explanations

This incorporates the feedback from Natasha.

* Ensure allowed users see the full name when limiting is enabled

There was a slight error when the code was moved to the
social_profile_privacy module that caused users with the "View full
name" permission to still see the nickname instead of the full name when
user name limiting was enabled.

* Issue #3039084 by chmez: Restore process for saving a list of fields which can be used in a profile.

* Issue #3039084 by chmez: Delete PHP class selector in calls in form alter hook.

* Issue #3039084 by chmez: Update main search index for adding nickname field.

* Issue #3039084 by chmez: Add nickname field to search processors.

* Issue #3039084 by chmez: Exclude first and last name fields of search index from search query when the limitation is enabled and userd does not have permission.

* Issue #3039084 by chmez: Add a nickname to username suggestions only when the nickname is not empty.

* Issue #3039084 by chmez: Fix deleting part of configuration when nickname field is creating.

* Issue #3043698 by Kingdutch: Search configuration for processors should not contain settings key

* Issue #3043714 by Kingdutch: Open Social should not index content immediately

* Reduce SocialProfileFieldOverride override impact size

We reduce the amount of configuration that we store in the override so
that we don't accidentally override something else.

* Simplify SocialProfileFieldOverride config overrides

We ensure that we don't copy existing configuration into the override.
We also revert a change that was made during the development of this
feature (#3039084) that referenced incorrect keys in the search_index
configuration. This was fixed in an earlier commit.

* Add return type suggestion for new hooks

* Handle disabling of first name or last name fields

The fact that first name or last name can be disabled individually means
that we need to do some special checking for all cases to ensure we
don't accidentally leak disabled information.

* Ensure search content is indexed in tests

In f9ff903 we decided that search content would no longer be indexed
immediately in order to improve Open Social performance. However, tests
rely on content being indexed during testing. To fix this a test step
has been added to trigger search indexing. This also allows the scenario
of content not being indexed to be tested should a user so desire.

* Fix config uninstall for social_profile_fields module

* Provide a test for the real name protection

* Fix display of names for name visibility permission

Instead of showing only the name instead of the nickname when a user is
allowed to view the name despite the user having a nickname we display
the full name in addition to the nickname.

This should only happen for users that have the permission however.

* Revert "Issue #3039084 by chmez: Exclude first and last name fields of search index from search query when the limitation is enabled and userd does not have permission."

This reverts commit 465d3c461570c31343e12125d2eb913714fab50b.

This moves back from excluding the fields from the search altogether to
adding additional constraints. This is needed because matches by a real
name only have to be excluded when the matched user has filled in a
nickname and no other fields matched.

* Rework logic for filtering searches

We only want to filter out queries that matched only on name and nothing
else when a nickname is filled.

* Add test scenario for username matches

* Add test for proper searching based on username

This also allows the restrict name search to be turned off.

* Fix case sensitivity breaking name search

* Add smokescreen test for username search

This ensures that our restricted username search test only fails if
there is an actual bug in the restricting code.

* Remove problematic edge case test for now

This will be fixed later. It can be fixed in a minor release but the
test shouldn't break the impeding major release.

* Trigger travis

* Remove problematic test step for now

See: https://www.drupal.org/project/social/issues/3047912

* 3039084 by kingdutch, jochemvn: Don't run subqueries when searchterm is not given