Issue #2898087 by Kingdutch: Remove the Secret Groups dependency on r4032login
There's no need to extend the r4032login exception handler. We can use
the power of the event system to turn the 403 exception into a 404
exception when it suits us.

This ensures that the presence of a secret group isn't leaked to users
that are not allowed to view the group.