Issue #2898087 by Kingdutch: Fix group visibility for non-members
Whether a user can see a group when they're not a member is now properly
decoupled from whether a user can create a secret group. This is
necessary because we may want to let all users create secret groups but
should not let all users see all of them. Non-sitemanager users should
only see those groups that they're a member of.