Issue #3050760 by bramtenhove: Ensure event max enroll limit takes anonymous enrollments into account
The enroll button for anonymous users is never showing the 'no spots left' disabled button. There was
only a check if the current user had an event enrollment. But when there was already one anonymous
enrollment, this would always be positive.

Also moved the code to a service. I didn't add dependency injection for the Event AN Enroll Form.